+++ CHANGELOG.md
@@ -36,6 +36,10 @@
 * Fixed an issue causing the traversed route line to be hidden after route completion when `NavigationMapView.traversedRouteColor` was provided.
 * Fixed an issue causing the traversed route line to reappear when switching to an alternative route and back.
 
+### Directions
+
+* Fixed a crash in case of unknown AttributeOptions usage.
+
 ## 3.19.0
 
 ### Packaging